requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
openai==1.3.0
httpx==0.25.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pytz==2023.3
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
psutil==5.9.6